+++ chaosk8s/daemonset/actions.py
@@ -1,4 +1,5 @@
 import json
+import logging
 import os.path
 
 import yaml
@@ -6,7 +7,6 @@ from chaoslib.exceptions import ActivityFailed
 from chaoslib.types import Secrets
 from kubernetes import client
 from kubernetes.client.rest import ApiException
-from logzero import logger
 
 from chaosk8s import create_k8s_api_client
 
@@ -15,6 +15,7 @@ __all__ = [
     "delete_daemon_set",
     "update_daemon_set",
 ]
+logger = logging.getLogger("chaostoolkit")
 
 
 def create_daemon_set(