+++ chaosk8s/daemonset/probes.py
@@ -1,9 +1,9 @@
+import logging
 from functools import partial
 
 import urllib3
 from chaoslib.types import Secrets
 from kubernetes import client, watch
-from logzero import logger
 
 from chaosk8s import create_k8s_api_client
 
@@ -13,6 +13,7 @@ __all__ = [
     "daemon_set_fully_available",
     "daemon_set_partially_available",
 ]
+logger = logging.getLogger("chaostoolkit")
 
 
 def daemon_set_available_and_healthy(