+++ chaosk8s/deployment/actions.py
@@ -1,5 +1,6 @@
 import datetime
 import json
+import logging
 import os.path
 
 import yaml
@@ -7,7 +8,6 @@ from chaoslib.exceptions import ActivityFailed
 from chaoslib.types import Secrets
 from kubernetes import client
 from kubernetes.client.rest import ApiException
-from logzero import logger
 
 from chaosk8s import create_k8s_api_client
 
@@ -17,6 +17,7 @@ __all__ = [
     "scale_deployment",
     "rollout_deployment",
 ]
+logger = logging.getLogger("chaostoolkit")
 
 
 def create_deployment(