+++ chaosk8s/pod/actions.py
@@ -1,5 +1,6 @@
 import datetime
 import json
+import logging
 import math
 import random
 import re
@@ -11,11 +12,11 @@ from chaoslib.types import Secrets
 from kubernetes import client, stream
 from kubernetes.client.models.v1_pod import V1Pod
 from kubernetes.stream.ws_client import ERROR_CHANNEL, STDOUT_CHANNEL
-from logzero import logger
 
 from chaosk8s import _log_deprecated, create_k8s_api_client
 
 __all__ = ["terminate_pods", "exec_in_pods"]
+logger = logging.getLogger("chaostoolkit")
 
 
 def terminate_pods(