+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Moved to PDM away from setuptools
 * Moved to Trusted Publisher to publish the package to Pypi
 * Removed black and isort dependencies to keep ruff
+* Dropped logzero dependency
 
 ## [0.35.0][] - 2024-03-14